server/requirements.txt
alembic==1.9.4
click==8.1.3
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.1
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
